Split iphone platform export template into multiple files
