+++ platform/iphone/export/godot_plugin_config.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  godot_plugin_config.h                                                */
+/*  godot_plugin_config.cpp                                              */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,92 +28,13 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef GODOT_PLUGIN_CONFIG_H
-#define GODOT_PLUGIN_CONFIG_H
+#include "godot_plugin_config.h"
 
-#include "core/error/error_list.h"
-#include "core/io/config_file.h"
-#include "core/string/ustring.h"
+#include "core/config/project_settings.h"
+#include "core/io/dir_access.h"
+#include "core/io/file_access.h"
 
-/*
- The `config` section and fields are required and defined as follow:
-- **name**: name of the plugin
-- **binary**: path to static `.a` library
-
-The `dependencies` and fields are optional.
-- **linked**: dependencies that should only be linked.
-- **embedded**: dependencies that should be linked and embedded into application.
-- **system**: system dependencies that should be linked.
-- **capabilities**: capabilities that would be used for `UIRequiredDeviceCapabilities` options in Info.plist file.
-- **files**: files that would be copied into application
-
-The `plist` section are optional.
-- **key**: key and value that would be added in Info.plist file.
- */
-
-struct PluginConfigIOS {
-	inline static const char *PLUGIN_CONFIG_EXT = ".gdip";
-
-	inline static const char *CONFIG_SECTION = "config";
-	inline static const char *CONFIG_NAME_KEY = "name";
-	inline static const char *CONFIG_BINARY_KEY = "binary";
-	inline static const char *CONFIG_INITIALIZE_KEY = "initialization";
-	inline static const char *CONFIG_DEINITIALIZE_KEY = "deinitialization";
-
-	inline static const char *DEPENDENCIES_SECTION = "dependencies";
-	inline static const char *DEPENDENCIES_LINKED_KEY = "linked";
-	inline static const char *DEPENDENCIES_EMBEDDED_KEY = "embedded";
-	inline static const char *DEPENDENCIES_SYSTEM_KEY = "system";
-	inline static const char *DEPENDENCIES_CAPABILITIES_KEY = "capabilities";
-	inline static const char *DEPENDENCIES_FILES_KEY = "files";
-	inline static const char *DEPENDENCIES_LINKER_FLAGS = "linker_flags";
-
-	inline static const char *PLIST_SECTION = "plist";
-
-	enum PlistItemType {
-		UNKNOWN,
-		STRING,
-		INTEGER,
-		BOOLEAN,
-		RAW,
-		STRING_INPUT,
-	};
-
-	struct PlistItem {
-		PlistItemType type;
-		String value;
-	};
-
-	// Set to true when the config file is properly loaded.
-	bool valid_config = false;
-	bool supports_targets = false;
-	// Unix timestamp of last change to this plugin.
-	uint64_t last_updated = 0;
-
-	// Required config section
-	String name;
-	String binary;
-	String initialization_method;
-	String deinitialization_method;
-
-	// Optional dependencies section
-	Vector<String> linked_dependencies;
-	Vector<String> embedded_dependencies;
-	Vector<String> system_dependencies;
-
-	Vector<String> files_to_copy;
-	Vector<String> capabilities;
-
-	Vector<String> linker_flags;
-
-	// Optional plist section
-	// String value is default value.
-	// Currently supports `string`, `boolean`, `integer`, `raw`, `string_input` types
-	// <name>:<type> = <value>
-	HashMap<String, PlistItem> plist;
-};
-
-static inline String resolve_local_dependency_path(String plugin_config_dir, String dependency_path) {
+String PluginConfigIOS::resolve_local_dependency_path(String plugin_config_dir, String dependency_path) {
 	String absolute_path;
 
 	if (dependency_path.is_empty()) {
@@ -130,7 +51,7 @@ static inline String resolve_local_dependency_path(String plugin_config_dir, Str
 	return absolute_path.replace(res_path, "res://");
 }
 
-static inline String resolve_system_dependency_path(String dependency_path) {
+String PluginConfigIOS::resolve_system_dependency_path(String dependency_path) {
 	String absolute_path;
 
 	if (dependency_path.is_empty()) {
@@ -146,7 +67,7 @@ static inline String resolve_system_dependency_path(String dependency_path) {
 	return system_path.plus_file(dependency_path);
 }
 
-static inline Vector<String> resolve_local_dependencies(String plugin_config_dir, Vector<String> p_paths) {
+Vector<String> PluginConfigIOS::resolve_local_dependencies(String plugin_config_dir, Vector<String> p_paths) {
 	Vector<String> paths;
 
 	for (int i = 0; i < p_paths.size(); i++) {
@@ -162,7 +83,7 @@ static inline Vector<String> resolve_local_dependencies(String plugin_config_dir
 	return paths;
 }
 
-static inline Vector<String> resolve_system_dependencies(Vector<String> p_paths) {
+Vector<String> PluginConfigIOS::resolve_system_dependencies(Vector<String> p_paths) {
 	Vector<String> paths;
 
 	for (int i = 0; i < p_paths.size(); i++) {
@@ -178,7 +99,7 @@ static inline Vector<String> resolve_system_dependencies(Vector<String> p_paths)
 	return paths;
 }
 
-static inline bool validate_plugin(PluginConfigIOS &plugin_config) {
+bool PluginConfigIOS::validate_plugin(PluginConfigIOS &plugin_config) {
 	bool valid_name = !plugin_config.name.is_empty();
 	bool valid_binary_name = !plugin_config.binary.is_empty();
 	bool valid_initialize = !plugin_config.initialization_method.is_empty();
@@ -213,7 +134,7 @@ static inline bool validate_plugin(PluginConfigIOS &plugin_config) {
 	return plugin_config.valid_config;
 }
 
-static inline String get_plugin_main_binary(PluginConfigIOS &plugin_config, bool p_debug) {
+String PluginConfigIOS::get_plugin_main_binary(PluginConfigIOS &plugin_config, bool p_debug) {
 	if (!plugin_config.supports_targets) {
 		return plugin_config.binary;
 	}
@@ -226,7 +147,7 @@ static inline String get_plugin_main_binary(PluginConfigIOS &plugin_config, bool
 	return plugin_binary_dir.plus_file(plugin_file);
 }
 
-static inline uint64_t get_plugin_modification_time(const PluginConfigIOS &plugin_config, const String &config_path) {
+uint64_t PluginConfigIOS::get_plugin_modification_time(const PluginConfigIOS &plugin_config, const String &config_path) {
 	uint64_t last_updated = FileAccess::get_modified_time(config_path);
 
 	if (!plugin_config.supports_targets) {
@@ -245,7 +166,7 @@ static inline uint64_t get_plugin_modification_time(const PluginConfigIOS &plugi
 	return last_updated;
 }
 
-static inline PluginConfigIOS load_plugin_config(Ref<ConfigFile> config_file, const String &path) {
+PluginConfigIOS PluginConfigIOS::load_plugin_config(Ref<ConfigFile> config_file, const String &path) {
 	PluginConfigIOS plugin_config = {};
 
 	if (!config_file.is_valid()) {
@@ -362,5 +283,3 @@ static inline PluginConfigIOS load_plugin_config(Ref<ConfigFile> config_file, co
 
 	return plugin_config;
 }
-
-#endif // GODOT_PLUGIN_CONFIG_H
